requirements.txt
opencv-python
tkcalendar
Pillow
python-dotenv
cmake
numpy
face-recognition==1.3.0
